Fix the y-axis flip issue.
Texture position and direction should be correct now.
We have to take care of gl's origin 0|0 sitting in the bottom left corner.
The GL code flips this within rendering to RenderBuffer and does a
framebuffer blit afterwards.
Within GLES we have to deal with y-inverted textures in the RenderBuffer command
and do the flip within rendering to output buffer (vdpau surface).
Therefore we also have to take care to invert the y coordinates.

Signed-off-by: Lucian Muresan <lucianm@users.sourceforge.net>